// src/renderer/extensions/file_based_loadorder/types/types.ts
export type LockedState = true | false | "true" | "false" | "always" | "never";
export type LoadOrder = ILoadOrderEntry[];
// ...
export interface ILoadOrderEntryExt extends ILoadOrderEntry {
  // The known index of this entry.
  index: number;

  // The fileId of the mod to which this LO entry belongs.
  //  There's currently no reliable way to determine the version of the mod
  //  due to BE data - the fileId is the only unique identifier we can use
  //  to detect a version change. (this is optional as some load order entries
  //  may be managed externally and not have a fileId)
  fileId?: number | undefined;
}

export interface ILoadOrderEntry<T = any> {
  // An arbitrary unique id for the load order item
  id: string;

  // Is this entry enabled ?
  enabled: boolean;

  // The entry's display name.
  name: string;

  // Is the entry locked into a certain position ?
  //  This only affects how the LO page renders the
  //  entry. The load order API will not enforce positions.
  locked?: LockedState;

  // The id of the mod to which this LO entry belongs.
  //  It's extremely important to set this property for entries
  //  generated from mods that are actively managed by Vortex; forgetting
  //  to do so can result in unexpected behaviour (such as entries not
  //  being included in a collection)
  //  can be left undefined if the entry is not managed by Vortex.
  modId?: string;

  // Custom data passed along with the load order entry
  data?: T;
}

export interface IInvalidResult {
  // The unique identifier of the load order entry that
  //  failed validation.
  id: string;

  // The reason why the load order entry is invalid.
  reason: string;
}

export interface IValidationResult {
  // Stores the invalid results (if any); the validation
  //  functor can provide a list of load order entry id's
  //  complemented with the reason why validation had failed
  //  which will be displayed to the user. The LO API will assume
  //  that the validation had passed successfully if no invalid
  //  results are provided.
  invalid: IInvalidResult[];
}
// ...
export interface ILoadOrderGameInfoExt extends ILoadOrderGameInfo {
  // The things I do to reduce complexity for extension developers...
  //  (and to block users from sending us reports which we can do nothing about)
  isContributed: boolean;
}

export class LoadOrderValidationError extends Error {
  private mValidationRes: IValidationResult;
  private mLoadOrder: LoadOrder;
  constructor(validationRes: IValidationResult, loadOrder: LoadOrder) {
    super("Invalid Load Order");
    this.name = "LoadOrderValidationError";
    this.mValidationRes = validationRes;
    this.mLoadOrder = loadOrder;
  }

  public get validationResult(): IValidationResult {
    return this.mValidationRes;
  }

  public get loadOrder(): LoadOrder {
    return this.mLoadOrder;
  }

  public get loadOrderEntryNames(): string {
    const lo = this.mLoadOrder
      .filter((entry) => !!entry)
      .map((entry) => entry.name)
      .join("\n");
    return lo;
  }
}